refactor: fastapi request body builder 경로 수정
